dxScheduler - rename Tooltip's files and move tests to __tests__ folder (#13809)
* Fix

* Revert h

* Refactor

* Fix lint
