+++ js/bundles/modules/parts/widgets-renovation.js
@@ -7,6 +7,6 @@ const renovation = require('../../../bundles/modules/renovation');
 renovation.dxrButton = require('../../../renovation/button.j').default;
 renovation.dxrWidget = require('../../../renovation/common/widget.j').default;
 renovation.dxrGridPager = require('../../../renovation/pager/grid_pager.j').default;
-renovation.dxrTooltipItemLayout = require('../../../renovation/scheduler/appointment-tooltip/item-layout.j').default;
+renovation.dxrTooltipItemLayout = require('../../../renovation/scheduler/appointment_tooltip/item_layout.j').default;
 /// BUNDLER_PARTS_END
 module.exports = renovation;
