+++ js/renovation/scheduler/appointment_tooltip/item_content.tsx
