+++ js/renovation/scheduler/appointment_tooltip/types.tsx
