+++ js/renovation/scheduler/appointment_tooltip/utils/default_functions.tsx
