+++ js/renovation/scheduler/appointment_tooltip/utils/get_current_appointment.tsx
